Move chrony specific code to model::chrony
